idempotency

+++ l1-contracts/bootstrap.sh
@@ -113,11 +113,15 @@ function release_git_push {
     git checkout -b "$branch_name"
   fi
 
-  git add .
-  git commit -m "Release $tag_name." >/dev/null
-  git tag -a "$tag_name" -m "Release $tag_name."
-  do_or_dryrun git push origin "$branch_name" --quiet
-  do_or_dryrun git push origin --quiet --force "$tag_name" --tags
+  if git rev-parse "$tag_name" >/dev/null 2>&1; then
+    echo "Tag $tag_name already exists. Skipping tag creation."
+  else
+    git add .
+    git commit -m "Release $tag_name." >/dev/null
+    git tag -a "$tag_name" -m "Release $tag_name."
+    do_or_dryrun git push origin "$branch_name" --quiet
+    do_or_dryrun git push origin --quiet --force "$tag_name" --tags
+  fi
 
   echo "Release complete ($tag_name) on branch $branch_name."
 }
